feat: add Context interface
To perform ontology saturation concurrently, it is necessary to group axioms into contexts that can be processed by workers independently. A context can only be processed by one thread at a time. If multiple axioms together can generate an inference, then these axioms must necessarily be contained in a common context. A context is active when its queue of to-do axioms is not empty. The operations that expose this interface must be thread-safe. Only 6 contexts can exist in this project. So this is a sealed interface.